backend/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
apify-client>=1.0.0
supabase>=2.0.0
firecrawl-py>=0.0.1
python-dateutil>=2.8.0
requests>=2.31.0
Pillow>=10.0.0

